handle tag and category delete

+++ app/Http/Controllers/CategoriesController.php
@@ -100,6 +100,10 @@ class CategoriesController extends Controller
      */
     public function destroy(Category $category)
     {
+        if ($category->posts->count() > 0) {
+            session()->flash('error', 'Category can not be deleted because it has some posts');
+            return redirect()->back();
+        }
         $category->delete();
 
         session()->flash('success', 'Category deleted successfully');
